+++ src/main/java/org/jsoup/helper/DataUtil.java
@@ -2,6 +2,7 @@ package org.jsoup.helper;
 
 import org.jsoup.UncheckedIOException;
 import org.jsoup.internal.ConstrainableInputStream;
+import org.jsoup.internal.Normalizer;
 import org.jsoup.internal.StringUtil;
 import org.jsoup.nodes.Comment;
 import org.jsoup.nodes.Document;
@@ -10,11 +11,12 @@ import org.jsoup.nodes.Node;
 import org.jsoup.nodes.XmlDeclaration;
 import org.jsoup.parser.Parser;
 import org.jsoup.select.Elements;
-//import org.jsoup.helper.BomCharset;
 
 import javax.annotation.Nullable;
 import java.io.BufferedReader;
 import java.io.CharArrayReader;
+import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -28,13 +30,14 @@ import java.util.Locale;
 import java.util.Random;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.zip.GZIPInputStream;
 
 /**
  * Internal static utilities for handling data.
  *
  */
 @SuppressWarnings("CharsetObjectCanBeUsed")
-public final class DataUtil {
+public final class DataUtil extends DocumentLoader {
     private static final Pattern charsetPattern = Pattern.compile("(?i)\\bcharset=\\s*(?:[\"'])?([^\\s,;\"']*)");
     public static final Charset UTF_8 = Charset.forName("UTF-8"); // Don't use StandardCharsets, as those only appear in Android API 19, and we target 10.
     static final String defaultCharsetName = UTF_8.name(); // used if not found in header or meta charset
@@ -45,6 +48,77 @@ public final class DataUtil {
     static final int boundaryLength = 32;
 
     private DataUtil() {}
+
+    /**
+     * Loads and parses a file to a Document, with the HtmlParser. Files that are compressed with gzip (and end in {@code .gz} or {@code .z})
+     * are supported in addition to uncompressed files.
+     *
+     * @param file file to load
+     * @param charsetName (optional) character set of input; specify {@code null} to attempt to autodetect. A BOM in
+     *     the file will always override this setting.
+     * @param baseUri base URI of document, to resolve relative links against
+     * @return Document
+     * @throws IOException on IO error
+     */
+//    public static Document load(File file, @Nullable String charsetName, String baseUri) throws IOException {
+//        return load(file, charsetName, baseUri, Parser.htmlParser());
+//    }
+//
+//    /**
+//     * Loads and parses a file to a Document. Files that are compressed with gzip (and end in {@code .gz} or {@code .z})
+//     * are supported in addition to uncompressed files.
+//     *
+//     * @param file file to load
+//     * @param charsetName (optional) character set of input; specify {@code null} to attempt to autodetect. A BOM in
+//     *     the file will always override this setting.
+//     * @param baseUri base URI of document, to resolve relative links against
+//     * @param parser alternate {@link Parser#xmlParser() parser} to use.
+//
+//     * @return Document
+//     * @throws IOException on IO error
+//     * @since 1.14.2
+//     */
+//    public static Document load(File file, @Nullable String charsetName, String baseUri, Parser parser) throws IOException {
+//        InputStream stream = new FileInputStream(file);
+//        String name = Normalizer.lowerCase(file.getName());
+//        if (name.endsWith(".gz") || name.endsWith(".z")) {
+//            // unfortunately file input streams don't support marks (why not?), so we will close and reopen after read
+//            boolean zipped;
+//            try {
+//                zipped = (stream.read() == 0x1f && stream.read() == 0x8b); // gzip magic bytes
+//            } finally {
+//                stream.close();
+//
+//            }
+//            stream = zipped ? new GZIPInputStream(new FileInputStream(file)) : new FileInputStream(file);
+//        }
+//        return parseInputStream(stream, charsetName, baseUri, parser);
+//    }
+//
+//    /**
+//     * Parses a Document from an input steam.
+//     * @param in input stream to parse. The stream will be closed after reading.
+//     * @param charsetName character set of input (optional)
+//     * @param baseUri base URI of document, to resolve relative links against
+//     * @return Document
+//     * @throws IOException on IO error
+//     */
+//    public static Document load(InputStream in, @Nullable String charsetName, String baseUri) throws IOException {
+//        return parseInputStream(in, charsetName, baseUri, Parser.htmlParser());
+//    }
+//
+//    /**
+//     * Parses a Document from an input steam, using the provided Parser.
+//     * @param in input stream to parse. The stream will be closed after reading.
+//     * @param charsetName character set of input (optional)
+//     * @param baseUri base URI of document, to resolve relative links against
+//     * @param parser alternate {@link Parser#xmlParser() parser} to use.
+//     * @return Document
+//     * @throws IOException on IO error
+//     */
+//    public static Document load(InputStream in, @Nullable String charsetName, String baseUri, Parser parser) throws IOException {
+//        return parseInputStream(in, charsetName, baseUri, parser);
+//    }
 
     /**
      * Writes the input stream to the output stream. Doesn't close them.
@@ -104,19 +178,8 @@ public final class DataUtil {
 
                 // look for <?xml encoding='ISO-8859-1'?>
                 if (foundCharset == null && doc.childNodeSize() > 0) {
-                    Node first = doc.childNode(0);
-                    XmlDeclaration decl = null;
-                    if (first instanceof XmlDeclaration)
-                        decl = (XmlDeclaration) first;
-                    else if (first instanceof Comment) {
-                        Comment comment = (Comment) first;
-                        if (comment.isXmlDeclaration())
-                            decl = comment.asXmlDeclaration();
-                    }
-                    if (decl != null) {
-                        if (decl.name().equalsIgnoreCase("xml"))
-                            foundCharset = decl.attr("encoding");
-                    }
+                    //implemented extracted method
+                    foundCharset = getString(doc, foundCharset);
                 }
                 foundCharset = validateCharset(foundCharset);
                 if (foundCharset != null && !foundCharset.equalsIgnoreCase(defaultCharsetName)) { // need to re-decode. (case insensitive check here to match how validate works)
@@ -160,6 +223,23 @@ public final class DataUtil {
             input.close();
         }
         return doc;
+    }
+
+    private static String getString(Document doc, String foundCharset) {
+        Node first = doc.childNode(0);
+        XmlDeclaration decl = null;
+        if (first instanceof XmlDeclaration)
+            decl = (XmlDeclaration) first;
+        else if (first instanceof Comment) {
+            Comment comment = (Comment) first;
+            if (comment.isXmlDeclaration())
+                decl = comment.asXmlDeclaration();
+        }
+        if (decl != null) {
+            if (decl.name().equalsIgnoreCase("xml"))
+                foundCharset = decl.attr("encoding");
+        }
+        return foundCharset;
     }
 
     /**
@@ -231,10 +311,10 @@ public final class DataUtil {
             buffer.rewind();
         }
         if (bom[0] == 0x00 && bom[1] == 0x00 && bom[2] == (byte) 0xFE && bom[3] == (byte) 0xFF || // BE
-            bom[0] == (byte) 0xFF && bom[1] == (byte) 0xFE && bom[2] == 0x00 && bom[3] == 0x00) { // LE
+                bom[0] == (byte) 0xFF && bom[1] == (byte) 0xFE && bom[2] == 0x00 && bom[3] == 0x00) { // LE
             return new BomCharset("UTF-32", false); // and I hope it's on your system
         } else if (bom[0] == (byte) 0xFE && bom[1] == (byte) 0xFF || // BE
-            bom[0] == (byte) 0xFF && bom[1] == (byte) 0xFE) {
+                bom[0] == (byte) 0xFF && bom[1] == (byte) 0xFE) {
             return new BomCharset("UTF-16", false); // in all Javas
         } else if (bom[0] == (byte) 0xEF && bom[1] == (byte) 0xBB && bom[2] == (byte) 0xBF) {
             return new BomCharset("UTF-8", true); // in all Javas
@@ -253,4 +333,3 @@ public final class DataUtil {
         }
     }
 }
-
