+++ src/main/java/org/jsoup/nodes/Comment.java
@@ -48,10 +48,10 @@ public class Comment extends LeafNode {
 
 //	void outerHtmlTail(Appendable accum, int depth, Document.OutputSettings out) {}
 //
-//    @Override
-//    public String toString() {
-//        return outerHtml();
-//    }
+    @Override
+    public String toString() {
+        return outerHtml();
+    }
 
     @Override
     public Comment clone() {
