+++ src/main/java/org/jsoup/nodes/Node.java
@@ -808,32 +808,32 @@ public abstract class Node implements Cloneable {
         return clone;
     }
 
-    private static class OuterHtmlVisitor implements NodeVisitor {
-        private final Appendable accum;
-        private final Document.OutputSettings out;
-
-        OuterHtmlVisitor(Appendable accum, Document.OutputSettings out) {
-            this.accum = accum;
-            this.out = out;
-            out.prepareEncoder();
-        }
-
-        public void head(Node node, int depth) {
-            try {
-				node.outerHtmlHead(accum, depth, out);
-			} catch (IOException exception) {
-				throw new SerializationException(exception);
-			}
-        }
-
-        public void tail(Node node, int depth) {
-            if (!node.nodeName().equals("#text")) { // saves a void hit.
-				try {
-					node.outerHtmlTail(accum, depth, out);
-				} catch (IOException exception) {
-					throw new SerializationException(exception);
-				}
-            }
-        }
-    }
+//    private static class OuterHtmlVisitor implements NodeVisitor {
+//        private final Appendable accum;
+//        private final Document.OutputSettings out;
+//
+//        OuterHtmlVisitor(Appendable accum, Document.OutputSettings out) {
+//            this.accum = accum;
+//            this.out = out;
+//            out.prepareEncoder();
+//        }
+//
+//        public void head(Node node, int depth) {
+//            try {
+//				node.outerHtmlHead(accum, depth, out);
+//			} catch (IOException exception) {
+//				throw new SerializationException(exception);
+//			}
+//        }
+//
+//        public void tail(Node node, int depth) {
+//            if (!node.nodeName().equals("#text")) { // saves a void hit.
+//				try {
+//					node.outerHtmlTail(accum, depth, out);
+//				} catch (IOException exception) {
+//					throw new SerializationException(exception);
+//				}
+//            }
+//        }
+//    }
 }
