+++ src/main/java/org/jsoup/select/AbstractEvaluator.java
@@ -1,8 +1,7 @@
 package org.jsoup.select;
 
 public abstract class AbstractEvaluator {
-
-      Evaluator getEvaluator(Evaluator currentEval, Evaluator newEval){
+      Evaluator  getEvaluator(char combinator, Evaluator currentEval, Evaluator newEval){
         return currentEval;
     };
 }
