+++ src/main/java/org/jsoup/select/QueryParser.java
@@ -4,7 +4,6 @@ import org.jsoup.internal.StringUtil;
 import org.jsoup.helper.Validate;
 import org.jsoup.parser.TokenQueue;
 
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
@@ -106,22 +105,38 @@ public class QueryParser {
         evals.clear();
 
         // for most combinators: change the current eval into an AND of the current eval and the new eval
-        currentEval = getEvaluator(combinator, currentEval, newEval);
+        switch (combinator) {
+            case '>':
+                currentEval = new CombiningEvaluator.And(new StructuralEvaluator.ImmediateParent(currentEval), newEval);
+                break;
+            case ' ':
+                currentEval = new CombiningEvaluator.And(new StructuralEvaluator.Parent(currentEval), newEval);
+                break;
+            case '+':
+                currentEval = new CombiningEvaluator.And(new StructuralEvaluator.ImmediatePreviousSibling(currentEval), newEval);
+                break;
+            case '~':
+                currentEval = new CombiningEvaluator.And(new StructuralEvaluator.PreviousSibling(currentEval), newEval);
+                break;
+            case ',':
+                CombiningEvaluator.Or or;
+                if (currentEval instanceof CombiningEvaluator.Or) {
+                    or = (CombiningEvaluator.Or) currentEval;
+                } else {
+                    or = new CombiningEvaluator.Or();
+                    or.add(currentEval);
+                }
+                or.add(newEval);
+                currentEval = or;
+                break;
+            default:
+                throw new Selector.SelectorParseException("Unknown combinator '%s'", combinator);
+        }
 
         if (replaceRightMost)
             ((CombiningEvaluator.Or) rootEval).replaceRightMostEvaluator(currentEval);
         else rootEval = currentEval;
         evals.add(rootEval);
-    }
-
-
-    private Evaluator getEvaluator(char combinator, Evaluator currentEval, Evaluator newEval) {
-        currentEval = EvaluatorFactory.createEvaluator(combinator);
-        if (currentEval == null) {
-            throw new Selector.SelectorParseException("Unknown combinator '%s'", combinator);
-        }
-        currentEval = currentEval.getEvaluator(currentEval, newEval);
-        return currentEval;
     }
 
 
@@ -182,33 +197,33 @@ public class QueryParser {
             matchesWholeText(true);
         else if (tq.matches(":not("))
             not();
-		else if (tq.matchChomp(":nth-child("))
-        	cssNthChild(false, false);
+        else if (tq.matchChomp(":nth-child("))
+            cssNthChild(false, false);
         else if (tq.matchChomp(":nth-last-child("))
-        	cssNthChild(true, false);
+            cssNthChild(true, false);
         else if (tq.matchChomp(":nth-of-type("))
-        	cssNthChild(false, true);
+            cssNthChild(false, true);
         else if (tq.matchChomp(":nth-last-of-type("))
-        	cssNthChild(true, true);
+            cssNthChild(true, true);
         else if (tq.matchChomp(":first-child"))
-        	evals.add(new Evaluator.IsFirstChild());
+            evals.add(new Evaluator.IsFirstChild());
         else if (tq.matchChomp(":last-child"))
-        	evals.add(new Evaluator.IsLastChild());
+            evals.add(new Evaluator.IsLastChild());
         else if (tq.matchChomp(":first-of-type"))
-        	evals.add(new Evaluator.IsFirstOfType());
+            evals.add(new Evaluator.IsFirstOfType());
         else if (tq.matchChomp(":last-of-type"))
-        	evals.add(new Evaluator.IsLastOfType());
+            evals.add(new Evaluator.IsLastOfType());
         else if (tq.matchChomp(":only-child"))
-        	evals.add(new Evaluator.IsOnlyChild());
+            evals.add(new Evaluator.IsOnlyChild());
         else if (tq.matchChomp(":only-of-type"))
-        	evals.add(new Evaluator.IsOnlyOfType());
+            evals.add(new Evaluator.IsOnlyOfType());
         else if (tq.matchChomp(":empty"))
-        	evals.add(new Evaluator.IsEmpty());
+            evals.add(new Evaluator.IsEmpty());
         else if (tq.matchChomp(":root"))
-        	evals.add(new Evaluator.IsRoot());
+            evals.add(new Evaluator.IsRoot());
         else if (tq.matchChomp(":matchText"))
             evals.add(new Evaluator.MatchText());
-		else // unhandled
+        else // unhandled
             throw new Selector.SelectorParseException("Could not parse query '%s': unexpected token at '%s'", query, tq.remainder());
 
     }
@@ -236,8 +251,8 @@ public class QueryParser {
         if (tagName.startsWith("*|")) {
             String plainTag = tagName.substring(2); // strip *|
             evals.add(new CombiningEvaluator.Or(
-                new Evaluator.Tag(plainTag),
-                new Evaluator.TagEndsWith(tagName.replace("*|", ":")))
+                    new Evaluator.Tag(plainTag),
+                    new Evaluator.TagEndsWith(tagName.replace("*|", ":")))
             );
         } else {
             // namespaces: if element name is "abc:def", selector must be "abc|def", so flip:
@@ -298,43 +313,43 @@ public class QueryParser {
     private void indexEquals() {
         evals.add(new Evaluator.IndexEquals(consumeIndex()));
     }
-    
+
     //pseudo selectors :first-child, :last-child, :nth-child, ...
     private static final Pattern NTH_AB = Pattern.compile("(([+-])?(\\d+)?)n(\\s*([+-])?\\s*\\d+)?", Pattern.CASE_INSENSITIVE);
     private static final Pattern NTH_B  = Pattern.compile("([+-])?(\\d+)");
 
-	private void cssNthChild(boolean backwards, boolean ofType) {
-		String argS = normalize(tq.chompTo(")"));
-		Matcher mAB = NTH_AB.matcher(argS);
-		Matcher mB = NTH_B.matcher(argS);
-		final int a, b;
-		if ("odd".equals(argS)) {
-			a = 2;
-			b = 1;
-		} else if ("even".equals(argS)) {
-			a = 2;
-			b = 0;
-		} else if (mAB.matches()) {
-			a = mAB.group(3) != null ? Integer.parseInt(mAB.group(1).replaceFirst("^\\+", "")) : 1;
-			b = mAB.group(4) != null ? Integer.parseInt(mAB.group(4).replaceFirst("^\\+", "")) : 0;
-		} else if (mB.matches()) {
-			a = 0;
-			b = Integer.parseInt(mB.group().replaceFirst("^\\+", ""));
-		} else {
-			throw new Selector.SelectorParseException("Could not parse nth-index '%s': unexpected format", argS);
-		}
-		if (ofType)
-			if (backwards)
-				evals.add(new Evaluator.IsNthLastOfType(a, b));
-			else
-				evals.add(new Evaluator.IsNthOfType(a, b));
-		else {
-			if (backwards)
-				evals.add(new Evaluator.IsNthLastChild(a, b));
-			else
-				evals.add(new Evaluator.IsNthChild(a, b));
-		}
-	}
+    private void cssNthChild(boolean backwards, boolean ofType) {
+        String argS = normalize(tq.chompTo(")"));
+        Matcher mAB = NTH_AB.matcher(argS);
+        Matcher mB = NTH_B.matcher(argS);
+        final int a, b;
+        if ("odd".equals(argS)) {
+            a = 2;
+            b = 1;
+        } else if ("even".equals(argS)) {
+            a = 2;
+            b = 0;
+        } else if (mAB.matches()) {
+            a = mAB.group(3) != null ? Integer.parseInt(mAB.group(1).replaceFirst("^\\+", "")) : 1;
+            b = mAB.group(4) != null ? Integer.parseInt(mAB.group(4).replaceFirst("^\\+", "")) : 0;
+        } else if (mB.matches()) {
+            a = 0;
+            b = Integer.parseInt(mB.group().replaceFirst("^\\+", ""));
+        } else {
+            throw new Selector.SelectorParseException("Could not parse nth-index '%s': unexpected format", argS);
+        }
+        if (ofType)
+            if (backwards)
+                evals.add(new Evaluator.IsNthLastOfType(a, b));
+            else
+                evals.add(new Evaluator.IsNthOfType(a, b));
+        else {
+            if (backwards)
+                evals.add(new Evaluator.IsNthLastChild(a, b));
+            else
+                evals.add(new Evaluator.IsNthChild(a, b));
+        }
+    }
 
     private int consumeIndex() {
         String indexS = tq.chompTo(")").trim();
@@ -357,8 +372,8 @@ public class QueryParser {
         String searchText = TokenQueue.unescape(tq.chompBalanced('(', ')'));
         Validate.notEmpty(searchText, query + "(text) query must not be empty");
         evals.add(own
-            ? new Evaluator.ContainsOwnText(searchText)
-            : new Evaluator.ContainsText(searchText));
+                ? new Evaluator.ContainsOwnText(searchText)
+                : new Evaluator.ContainsText(searchText));
     }
 
     private void containsWholeText(boolean own) {
@@ -367,8 +382,8 @@ public class QueryParser {
         String searchText = TokenQueue.unescape(tq.chompBalanced('(', ')'));
         Validate.notEmpty(searchText, query + "(text) query must not be empty");
         evals.add(own
-            ? new Evaluator.ContainsWholeOwnText(searchText)
-            : new Evaluator.ContainsWholeText(searchText));
+                ? new Evaluator.ContainsWholeOwnText(searchText)
+                : new Evaluator.ContainsWholeText(searchText));
     }
 
     // pseudo selector :containsData(data)
@@ -387,8 +402,8 @@ public class QueryParser {
         Validate.notEmpty(regex, query + "(regex) query must not be empty");
 
         evals.add(own
-            ? new Evaluator.MatchesOwn(Pattern.compile(regex))
-            : new Evaluator.Matches(Pattern.compile(regex)));
+                ? new Evaluator.MatchesOwn(Pattern.compile(regex))
+                : new Evaluator.Matches(Pattern.compile(regex)));
     }
 
     // :matches(regex), matchesOwn(regex)
@@ -399,8 +414,8 @@ public class QueryParser {
         Validate.notEmpty(regex, query + "(regex) query must not be empty");
 
         evals.add(own
-            ? new Evaluator.MatchesWholeOwnText(Pattern.compile(regex))
-            : new Evaluator.MatchesWholeText(Pattern.compile(regex)));
+                ? new Evaluator.MatchesWholeOwnText(Pattern.compile(regex))
+                : new Evaluator.MatchesWholeText(Pattern.compile(regex)));
     }
 
     // :not(selector)
